Added incursions, area calculation, area filling.
Incursions are done by making exactly 2 directional changes during its creation, and resolved by heading back to a outer border. This ensures that a rectangle is always made.

The area of this rectangle is then calculated, compared to the MAP_AREA and a percentage is given that the player occupies.

If the Qix is inside the rectangle when formed, the game will assume the player trapped it and as a result the player claims all the area outside of the rectangle as the Qix resides within the incursion.